pyproject.toml
[project]
name = "ai-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.70.0",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.9.0",
]
